requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
django-decouple==2.1
djangorestframework==3.15.2
drf-yasg==1.21.8
gunicorn==23.0.0
inflection==0.5.1
packaging==24.2
pillow==11.1.0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
whitenoise==6.8.2
